@@ -17,12 +17,10 @@ use League\Flysystem\DirectoryAttributes;
 use League\Flysystem\PathPrefixer;
 use League\Flysystem\Visibility;
 use OSS\Core\OssException;
+use OSS\Credentials\StaticCredentialsProvider;
 use OSS\OssClient;
 use League\Flysystem\FileAttributes;
 use League\Flysystem\FilesystemAdapter;
-use League\Flysystem\FilesystemException;
-use League\Flysystem\InvalidVisibilityProvided;
-use League\Flysystem\StorageAttributes;
 use League\Flysystem\UnableToCopyFile;
 use League\Flysystem\UnableToCreateDirectory;
 use League\Flysystem\UnableToDeleteDirectory;
@@ -42,11 +40,10 @@ class OssAdapter implements FilesystemAdapter
 {
     use SignatureTrait;
 
-    // 系统参数
     /**
-     *
+     * 系统参数
      */
-    const SYSTEM_FIELD = [
+    public const SYSTEM_FIELD = [
         'bucket'   => '${bucket}',
         'etag'     => '${etag}',
         'filename' => '${object}',
@@ -57,35 +54,25 @@ class OssAdapter implements FilesystemAdapter
         'format'   => '${imageInfo.format}',
     ];
 
-    /**
-     * @var
-     */
     protected $accessKeyId;
 
-    /**
-     * @var
-     */
     protected $accessKeySecret;
 
-    /**
-     * @var
-     */
     protected $endpoint;
 
-    /**
-     * @var
-     */
-    protected $bucket;
+    protected string $bucketName = '';
+
+    protected bool $isCName = false;
 
     /**
-     * @var
+     * @var array<string, array>
      */
-    protected $isCName;
+    protected array $buckets = [];
 
     /**
-     * @var array
+     * @var array<string, OssAdapter>
      */
-    protected $buckets;
+    protected array $bucketAdapters = [];
 
     /**
      * @var OssClient
@@ -93,111 +80,148 @@ class OssAdapter implements FilesystemAdapter
     protected $client;
 
     /**
-     * @var array|mixed[]
+     * @var array<string, mixed>
      */
-    protected $params;
+    protected array $params = [];
 
     /**
      * @var bool
      */
-    protected $useSSL = false;
+    protected bool $useSSL = false;
 
     /**
      * @var string|null
      */
-    protected $cdnUrl = null;
+    protected ?string $cdnUrl = null;
 
     /**
      * @var PathPrefixer
      */
     protected $prefixer;
 
-    /**
-     * @param        $accessKeyId
-     * @param        $accessKeySecret
-     * @param        $endpoint
-     * @param        $bucket
-     * @param bool   $isCName
-     * @param string $prefix
-     * @param array  $buckets
-     * @param        ...$params
-     *
-     * @throws OssException
-     */
     public function __construct($accessKeyId, $accessKeySecret, $endpoint, $bucket, bool $isCName = false, string $prefix = '', array $buckets = [], ...$params)
     {
         $this->accessKeyId = $accessKeyId;
         $this->accessKeySecret = $accessKeySecret;
         $this->endpoint = $endpoint;
-        $this->bucket = $bucket;
+        $this->bucketName = $bucket;
         $this->isCName = $isCName;
         $this->prefixer = new PathPrefixer($prefix, DIRECTORY_SEPARATOR);
         $this->buckets = $buckets;
         $this->params = $params;
-        $this->initClient();
-        $this->checkEndpoint();
+        $this->initDefaultBucketAdapter();
+    }
+
+    public function getBucketName(): string
+    {
+        return $this->bucketName;
     }
 
     /**
-     * 设置cdn的url.
+     * init default bucket adapter.
+     *
+     * @return \Iidestiny\Flysystem\Oss\OssAdapter
+     */
+    protected function initDefaultBucketAdapter(): OssAdapter
+    {
+        $this->initClient()
+            ->checkEndpoint()
+            ->bucketAdapters[$this->bucketName] = $this;
+
+        return $this;
+    }
+
+    /**
+     * set cdn url.
      *
      * @param string|null $url
+     * @return $this
      */
-    public function setCdnUrl(?string $url)
+    public function setCdnUrl(?string $url): OssAdapter
     {
         $this->cdnUrl = $url;
+
+        return $this;
     }
 
-    /**
-     * @return OssClient
-     */
     public function ossKernel(): OssClient
     {
         return $this->getClient();
     }
 
     /**
-     * 调用不同的桶配置.
-     *
-     * @param $bucket
+     * get bucket adapter by bucket name.
      *
      * @return $this
-     * @throws OssException
-     * @throws \Exception
+     * @throws \InvalidArgumentException
      */
     public function bucket($bucket): OssAdapter
     {
+        return $this->bucketAdapters[$bucket] ?? ($this->bucketAdapters[$bucket] = $this->createBucketAdapter($bucket));
+    }
+
+    /**
+     * create bucket adapter by bucket name.
+     *
+     * @param string $bucket
+     * @return \Iidestiny\Flysystem\Oss\OssAdapter
+     * @throws \InvalidArgumentException
+     */
+    protected function createBucketAdapter(string $bucket): OssAdapter
+    {
         if (!isset($this->buckets[$bucket])) {
-            throw new \Exception('bucket is not exist.');
+            throw new \InvalidArgumentException(sprintf('Bucket "%s" does not exist.', $bucket));
         }
-        $bucketConfig = $this->buckets[$bucket];
 
-        $this->accessKeyId = $bucketConfig['access_key'];
-        $this->accessKeySecret = $bucketConfig['secret_key'];
-        $this->endpoint = $bucketConfig['endpoint'];
-        $this->bucket = $bucketConfig['bucket'];
-        $this->isCName = $bucketConfig['isCName'];
+        $config = $this->buckets[$bucket];
+        $extra = array_merge($this->params, $this->extraConfig($config));
 
-        $this->initClient();
-        $this->checkEndpoint();
+        // new bucket adapter
+        $adapter = new self(
+            $config['access_key'] ?? $this->accessKeyId,
+            $config['secret_key'] ?? $this->accessKeySecret,
+            $config['endpoint'] ?? $this->endpoint,
+            $config['bucket'],
+            $config['isCName'] ?? false,
+            $config['root'] ?? '',
+            [],
+            ...$extra);
 
-        return $this;
+        return $adapter->setCdnUrl($config['url'] ?? null)->initDefaultBucketAdapter();
+    }
+
+    /**
+     * extract extra config.
+     *
+     * @param array $config
+     * @return array
+     */
+    protected function extraConfig(array $config): array
+    {
+        return array_diff_key($config, array_flip(['driver', 'root', 'buckets', 'access_key', 'secret_key',
+            'endpoint', 'bucket', 'isCName', 'url']));
     }
 
     /**
      * init oss client.
      *
-     * @throws OssException
+     * @return $this
      */
-    protected function initClient()
+    protected function initClient(): OssAdapter
     {
-        $this->client = new OssClient($this->accessKeyId, $this->accessKeySecret, $this->endpoint, $this->isCName, ...$this->params);
+        $provider = new StaticCredentialsProvider($this->accessKeyId, $this->accessKeySecret, $this->params['securityToken'] ?? null);
+
+        $this->client = new OssClient(array_merge([
+            'endpoint' => rtrim($this->endpoint, '/'),
+            'cname'    => $this->isCName,
+            'provider' => $provider,
+        ], $this->params));
+
+        return $this;
     }
 
     /**
      * get ali sdk kernel class.
-     *
-     * @return OssClient
      */
     public function getClient(): OssClient
     {
@@ -206,25 +230,14 @@ class OssAdapter implements FilesystemAdapter
 
     /**
      * 验签.
-     *
-     * @return array
      */
     public function verify(): array
     {
-        // oss 前面header、公钥 header
-        $authorizationBase64 = '';
-        $pubKeyUrlBase64 = '';
-
-        if (isset($_SERVER['HTTP_AUTHORIZATION'])) {
-            $authorizationBase64 = $_SERVER['HTTP_AUTHORIZATION'];
-        }
-
-        if (isset($_SERVER['HTTP_X_OSS_PUB_KEY_URL'])) {
-            $pubKeyUrlBase64 = $_SERVER['HTTP_X_OSS_PUB_KEY_URL'];
-        }
-
+        // oss 前面 header、公钥 header
+        $authorizationBase64 = $_SERVER['HTTP_AUTHORIZATION'] ?? '';
+        $pubKeyUrlBase64 = $_SERVER['HTTP_X_OSS_PUB_KEY_URL'] ?? '';
         // 验证失败
-        if ('' == $authorizationBase64 || '' == $pubKeyUrlBase64) {
+        if (empty($authorizationBase64) || empty($pubKeyUrlBase64)) {
             return [false, ['CallbackFailed' => 'authorization or pubKeyUrl is null']];
         }
 
@@ -237,9 +250,7 @@ class OssAdapter implements FilesystemAdapter
         curl_setopt($ch, CURLOPT_URL, $pubKeyUrl);
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
         curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, 10);
-        $pubKey = curl_exec($ch);
-
-        if ('' == $pubKey) {
+        if (!$pubKey = curl_exec($ch)) {
             return [false, ['CallbackFailed' => 'curl is fail']];
         }
 
@@ -257,10 +268,13 @@ class OssAdapter implements FilesystemAdapter
         $ok = openssl_verify($authStr, $authorization, $pubKey, OPENSSL_ALGO_MD5);
 
         if (1 !== $ok) {
+            curl_close($ch);
+
             return [false, ['CallbackFailed' => 'verify is fail, Illegal data']];
         }
 
         parse_str($body, $data);
+        curl_close($ch);
 
         return [true, $data];
     }
@@ -268,18 +282,18 @@ class OssAdapter implements FilesystemAdapter
     /**
      * oss 直传配置.
      *
-     * @param string $prefix
-     * @param null   $callBackUrl
-     * @param array  $customData
-     * @param int    $expire
-     * @param int    $contentLengthRangeValue
-     * @param array  $systemData
-     *
-     * @return false|string
-     *
-     * @throws \Exception
+     * @param string $prefix    目录前缀
+     * @param null $callBackUrl 回调地址
+     * @param array $customData 自定义参数
+     * @param int $expire       过期时间（秒）
+     * @param array $systemData 系统接收参数，回调时会返回
+     * @param array $policyData 自定义 policy 参数
+     *                          see: https://help.aliyun.com/zh/oss/developer-reference/postobject#section-d5z-1ww-wdb
+     * @return string
+     * @throws \JsonException|\InvalidArgumentException|\DateMalformedStringException
+     * @see https://help.aliyun.com/zh/oss/use-cases/overview-20
      */
-    public function signatureConfig(string $prefix = '', $callBackUrl = null, array $customData = [], int $expire = 30, int $contentLengthRangeValue = 1048576000, array $systemData = [])
+    public function signatureConfig(string $prefix = '', $callBackUrl = null, array $customData = [], int $expire = 30, array $systemData = [], array $policyData = []): string
     {
         $prefix = $this->prefixer->prefixPath($prefix);
 
@@ -289,8 +303,8 @@ class OssAdapter implements FilesystemAdapter
             $system = self::SYSTEM_FIELD;
         } else {
             foreach ($systemData as $key => $value) {
-                if (!in_array($value, self::SYSTEM_FIELD)) {
-                    throw new \InvalidArgumentException("Invalid oss system filed: ${value}");
+                if (!in_array($value, self::SYSTEM_FIELD, true)) {
+                    throw new \InvalidArgumentException(sprintf('Invalid oss system filed: %s', $value));
                 }
                 $system[$key] = $value;
             }
@@ -311,33 +325,41 @@ class OssAdapter implements FilesystemAdapter
             'callbackBody'     => urldecode(http_build_query(array_merge($system, $data))),
             'callbackBodyType' => 'application/x-www-form-urlencoded',
         ];
-        $callbackString = json_encode($callbackParam);
+        $callbackString = json_encode($callbackParam, JSON_THROW_ON_ERROR);
         $base64CallbackBody = base64_encode($callbackString);
 
         $now = time();
         $end = $now + $expire;
         $expiration = $this->gmt_iso8601($end);
 
-        // 最大文件大小.用户可以自己设置
-        $condition = [
-            0 => 'content-length-range',
-            1 => 0,
-            2 => $contentLengthRangeValue,
-        ];
-        $conditions[] = $condition;
-
-        $start = [
+        // 如果用户没有设置文件大小，需要设置默认值
+        $hasContentLengthRange = false;
+        $contentLengthRangeKey = 'content-length-range';
+        foreach ($policyData as $item) {
+            if (isset($item[0]) && $item[0] === $contentLengthRangeKey) {
+                $hasContentLengthRange = true;
+                break;
+            }
+        }
+        if (!$hasContentLengthRange) {
+            $condition = [
+                0 => $contentLengthRangeKey,
+                1 => 0, // min: 0
+                2 => 1048576000, // max: 1GB
+            ];
+            $conditions[] = $condition;
+        }
+        $conditions[] = [
             0 => 'starts-with',
             1 => '$key',
             2 => $prefix,
         ];
-        $conditions[] = $start;
 
         $arr = [
             'expiration' => $expiration,
-            'conditions' => $conditions,
+            'conditions' => array_merge($conditions, $policyData), // 将自定义policy参数一起合并
         ];
-        $policy = json_encode($arr);
+        $policy = json_encode($arr, JSON_THROW_ON_ERROR);
         $base64Policy = base64_encode($policy);
         $stringToSign = $base64Policy;
         $signature = base64_encode(hash_hmac('sha1', $stringToSign, $this->accessKeySecret, true));
@@ -352,46 +374,34 @@ class OssAdapter implements FilesystemAdapter
         $response['callback-var'] = $callbackVar;
         $response['dir'] = $prefix;  // 这个参数是设置用户上传文件时指定的前缀。
 
-        return json_encode($response);
+        return json_encode($response, JSON_THROW_ON_ERROR);
     }
 
     /**
-     * sign url
+     * sign url.
      *
-     * @param        $path
-     * @param        $timeout
-     * @param array  $options
-     * @param string $method
-     *
-     * @return false|\OSS\Http\ResponseCore|string
+     * @return false|string
      */
-    public function getTemporaryUrl($path, $timeout, array $options = [], string $method = OssClient::OSS_HTTP_GET)
+    public function getTemporaryUrl(string $path, int $timeout, array $options = [], string $method = OssClient::OSS_HTTP_GET): bool|string
     {
         $path = $this->prefixer->prefixPath($path);
 
         try {
-            $path = $this->client->signUrl($this->bucket, $path, $timeout, $method, $options);
-        } catch (OssException $exception) {
+            $path = $this->client->signUrl($this->bucketName, $path, $timeout, $method, $options);
+        } catch (OssException $ossException) {
             return false;
         }
 
         return $path;
     }
 
-    /**
-     * @param string $path
-     * @param string $contents
-     * @param Config $config
-     *
-     * @return void
-     */
     public function write(string $path, string $contents, Config $config): void
     {
         $path = $this->prefixer->prefixPath($path);
         $options = $config->get('options', []);
 
         try {
-            $this->client->putObject($this->bucket, $path, $contents, $options);
+            $this->client->putObject($this->bucketName, $path, $contents, $options);
         } catch (\Exception $exception) {
             throw UnableToWriteFile::atLocation($path, $exception->getMessage());
         }
@@ -400,11 +410,7 @@ class OssAdapter implements FilesystemAdapter
     /**
      * Write a new file using a stream.
      *
-     * @param string $path
-     * @param        $contents
-     * @param Config $config
-     *
-     * @return void
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function writeStream(string $path, $contents, Config $config): void
     {
@@ -412,19 +418,12 @@ class OssAdapter implements FilesystemAdapter
         $options = $config->get('options', []);
 
         try {
-            $this->client->uploadStream($this->bucket, $this->prefixer->prefixPath($path), $contents, $options);
+            $this->client->uploadStream($this->bucketName, $path, $contents, $options);
         } catch (OssException $exception) {
             throw UnableToWriteFile::atLocation($path, $exception->getErrorCode(), $exception);
         }
     }
 
-    /**
-     * @param string $source
-     * @param string $destination
-     * @param Config $config
-     *
-     * @return void
-     */
     public function move(string $source, string $destination, Config $config): void
     {
         try {
@@ -436,11 +435,7 @@ class OssAdapter implements FilesystemAdapter
     }
 
     /**
-     * @param string $source
-     * @param string $destination
-     * @param Config $config
-     *
-     * @return void
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function copy(string $source, string $destination, Config $config): void
     {
@@ -448,8 +443,8 @@ class OssAdapter implements FilesystemAdapter
         $newPath = $this->prefixer->prefixPath($destination);
 
         try {
-            $this->client->copyObject($this->bucket, $path, $this->bucket, $newPath);
-        } catch (OssException $exception) {
+            $this->client->copyObject($this->bucketName, $path, $this->bucketName, $newPath);
+        } catch (OssException $ossException) {
             throw UnableToCopyFile::fromLocationTo($source, $destination);
         }
     }
@@ -457,26 +452,22 @@ class OssAdapter implements FilesystemAdapter
     /**
      * delete a file.
      *
-     * @param string $path
-     *
-     * @return void
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function delete(string $path): void
     {
         $path = $this->prefixer->prefixPath($path);
 
         try {
-            $this->client->deleteObject($this->bucket, $path);
+            $this->client->deleteObject($this->bucketName, $path);
         } catch (OssException $ossException) {
             throw UnableToDeleteFile::atLocation($path);
         }
     }
 
     /**
-     * @param string $path
-     *
-     * @return void
-     * @throws OssException
+     * @throws OssException|\OSS\Http\RequestCore_Exception
+     * @throws \Exception
      */
     public function deleteDirectory(string $path): void
     {
@@ -489,28 +480,25 @@ class OssAdapter implements FilesystemAdapter
                     continue;
                 }
                 $files[] = $this->prefixer->prefixPath($content->path());
-                if ($i && $i % 100 == 0) {
-                    $this->client->deleteObjects($this->bucket, $files);
+                if ($i && 0 === $i % 100) {
+                    $this->client->deleteObjects($this->bucketName, $files);
                     $files = [];
                 }
             }
-            !empty($files) && $this->client->deleteObjects($this->bucket, $files);
-            $this->client->deleteObject($this->bucket, $this->prefixer->prefixDirectoryPath($path));
+            !empty($files) && $this->client->deleteObjects($this->bucketName, $files);
+            $this->client->deleteObject($this->bucketName, $this->prefixer->prefixDirectoryPath($path));
         } catch (OssException $exception) {
             throw UnableToDeleteDirectory::atLocation($path, $exception->getErrorCode(), $exception);
         }
     }
 
     /**
-     * @param string $path
-     * @param Config $config
-     *
-     * @return void
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function createDirectory(string $path, Config $config): void
     {
         try {
-            $this->client->createObjectDir($this->bucket, $this->prefixer->prefixPath($path));
+            $this->client->createObjectDir($this->bucketName, $this->prefixer->prefixPath($path));
         } catch (OssException $exception) {
             throw UnableToCreateDirectory::dueToFailure($path, $exception);
         }
@@ -521,8 +509,7 @@ class OssAdapter implements FilesystemAdapter
      *
      * @param string $path
      * @param string $visibility
-     *
-     * @return array|bool|false
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function setVisibility(string $path, string $visibility): void
     {
@@ -531,48 +518,52 @@ class OssAdapter implements FilesystemAdapter
         $acl = Visibility::PUBLIC === $visibility ? OssClient::OSS_ACL_TYPE_PUBLIC_READ : OssClient::OSS_ACL_TYPE_PRIVATE;
 
         try {
-            $this->client->putObjectAcl($this->bucket, $object, $acl);
+            $this->client->putObjectAcl($this->bucketName, $object, $acl);
         } catch (OssException $exception) {
             throw UnableToSetVisibility::atLocation($path, $exception->getMessage());
         }
     }
 
     /**
-     * @param string $path
-     *
-     * @return FileAttributes
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function visibility(string $path): FileAttributes
     {
         try {
-            $acl = $this->client->getObjectAcl($this->bucket, $this->prefixer->prefixPath($path), []);
+            $acl = $this->client->getObjectAcl($this->bucketName, $this->prefixer->prefixPath($path), []);
         } catch (OssException $exception) {
             throw UnableToRetrieveMetadata::visibility($path, $exception->getMessage());
         }
 
-        return new FileAttributes($path, null, $acl === OssClient::OSS_ACL_TYPE_PRIVATE ? Visibility::PRIVATE : Visibility::PUBLIC);
+        return new FileAttributes($path, null, OssClient::OSS_ACL_TYPE_PRIVATE === $acl ? Visibility::PRIVATE : Visibility::PUBLIC);
     }
 
     /**
      * Check whether a file exists.
      *
      * @param string $path
-     *
-     * @return array|bool|null
+     * @return bool
+     * @throws \OSS\Core\OssException
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function fileExists(string $path): bool
     {
         $path = $this->prefixer->prefixPath($path);
 
-        return $this->client->doesObjectExist($this->bucket, $path);
+        return $this->client->doesObjectExist($this->bucketName, $path);
+    }
+
+    /**
+     * @throws \OSS\Core\OssException
+     * @throws \OSS\Http\RequestCore_Exception
+     */
+    public function directoryExists(string $path): bool
+    {
+        return $this->client->doesObjectExist($this->bucketName, $this->prefixer->prefixDirectoryPath($path));
     }
 
     /**
      * Get resource url.
-     *
-     * @param string $path
-     *
-     * @return string
      */
     public function getUrl(string $path): string
     {
@@ -587,17 +578,13 @@ class OssAdapter implements FilesystemAdapter
 
     /**
      * read a file.
-     *
-     * @param string $path
-     *
-     * @return string
      */
     public function read(string $path): string
     {
         $path = $this->prefixer->prefixPath($path);
 
         try {
-            return $this->client->getObject($this->bucket, $path);
+            return $this->client->getObject($this->bucketName, $path);
         } catch (\Exception $exception) {
             throw UnableToReadFile::fromLocation($path, $exception->getMessage());
         }
@@ -606,16 +593,15 @@ class OssAdapter implements FilesystemAdapter
     /**
      * read a file stream.
      *
-     * @param string $path
-     *
-     * @return array|bool|false
+     * @return bool|resource
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function readStream(string $path)
     {
         $stream = fopen('php://temp', 'w+b');
 
         try {
-            fwrite($stream, $this->client->getObject($this->bucket, $path, [OssClient::OSS_FILE_DOWNLOAD => $stream]));
+            fwrite($stream, $this->client->getObject($this->bucketName, $path, [OssClient::OSS_FILE_DOWNLOAD => $stream]));
         } catch (OssException $exception) {
             fclose($stream);
             throw UnableToReadFile::fromLocation($path, $exception->getMessage());
@@ -626,10 +612,6 @@ class OssAdapter implements FilesystemAdapter
     }
 
     /**
-     * @param string $path
-     * @param bool   $deep
-     *
-     * @return iterable
      * @throws \Exception
      */
     public function listContents(string $path, bool $deep): iterable
@@ -643,7 +625,7 @@ class OssAdapter implements FilesystemAdapter
             ];
 
             try {
-                $listObjectInfo = $this->client->listObjects($this->bucket, $options);
+                $listObjectInfo = $this->client->listObjects($this->bucketName, $options);
                 $nextMarker = $listObjectInfo->getNextMarker();
             } catch (OssException $exception) {
                 throw new \Exception($exception->getErrorMessage(), 0, $exception);
@@ -656,8 +638,8 @@ class OssAdapter implements FilesystemAdapter
                     continue;
                 }
                 yield new DirectoryAttributes($subPath);
-                if ($deep === true) {
-                    $contents = $this->listContents($subPath, $deep);
+                if ($deep) {
+                    $contents = $this->listContents($subPath, true);
                     foreach ($contents as $content) {
                         yield $content;
                     }
@@ -669,35 +651,33 @@ class OssAdapter implements FilesystemAdapter
                 foreach ($listObject as $objectInfo) {
                     $objectPath = $this->prefixer->stripPrefix($objectInfo->getKey());
                     $objectLastModified = strtotime($objectInfo->getLastModified());
-                    if (substr($objectPath, -1, 1) == '/') {
+                    if ('/' == substr($objectPath, -1, 1)) {
                         continue;
                     }
                     yield new FileAttributes($objectPath, $objectInfo->getSize(), null, $objectLastModified);
                 }
             }
 
-            if ($listObjectInfo->getIsTruncated() !== "true") {
+            if ('true' !== $listObjectInfo->getIsTruncated()) {
                 break;
             }
         }
     }
 
     /**
-     * @param $path
-     *
-     * @return FileAttributes
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function getMetadata($path): FileAttributes
     {
         try {
-            $result = $this->client->getObjectMeta($this->bucket, $this->prefixer->prefixPath($path));
+            $result = $this->client->getObjectMeta($this->bucketName, $this->prefixer->prefixPath($path));
         } catch (OssException $exception) {
-            throw UnableToRetrieveMetadata::create($path, "metadata", $exception->getErrorCode(), $exception);
+            throw UnableToRetrieveMetadata::create($path, 'metadata', $exception->getErrorCode(), $exception);
         }
 
-        $size = isset($result["content-length"]) ? intval($result["content-length"]) : 0;
-        $timestamp = isset($result["last-modified"]) ? strtotime($result["last-modified"]) : 0;
-        $mimetype = $result["content-type"] ?? "";
+        $size = (int)($result['content-length'] ?? 0);
+        $timestamp = isset($result['last-modified']) ? strtotime($result['last-modified']) : 0;
+        $mimetype = $result['content-type'] ?? '';
 
         return new FileAttributes($path, $size, null, $timestamp, $mimetype);
     }
@@ -706,15 +686,16 @@ class OssAdapter implements FilesystemAdapter
      * get the size of file.
      *
      * @param string $path
-     *
-     * @return array|false
+     * @return \League\Flysystem\FileAttributes
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function fileSize(string $path): FileAttributes
     {
         $meta = $this->getMetadata($path);
-        if ($meta->fileSize() === null) {
+        if (null === $meta->fileSize()) {
             throw UnableToRetrieveMetadata::fileSize($path);
         }
+
         return $meta;
     }
 
@@ -722,13 +703,13 @@ class OssAdapter implements FilesystemAdapter
      * get mime type.
      *
      * @param string $path
-     *
-     * @return array|false
+     * @return \League\Flysystem\FileAttributes
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function mimeType(string $path): FileAttributes
     {
         $meta = $this->getMetadata($path);
-        if ($meta->mimeType() === null) {
+        if (null === $meta->mimeType()) {
             throw UnableToRetrieveMetadata::mimeType($path);
         }
 
@@ -739,51 +720,54 @@ class OssAdapter implements FilesystemAdapter
      * get timestamp.
      *
      * @param string $path
-     *
-     * @return array|false
+     * @return \League\Flysystem\FileAttributes
+     * @throws \OSS\Http\RequestCore_Exception
      */
     public function lastModified(string $path): FileAttributes
     {
         $meta = $this->getMetadata($path);
-        if ($meta->lastModified() === null) {
+        if (null === $meta->lastModified()) {
             throw UnableToRetrieveMetadata::lastModified($path);
         }
+
         return $meta;
     }
 
     /**
      * normalize Host.
-     *
-     * @return string
      */
     protected function normalizeHost(): string
     {
         if ($this->isCName) {
             $domain = $this->endpoint;
         } else {
-            $domain = $this->bucket . '.' . $this->endpoint;
+            $domain = $this->bucketName . '.' . $this->endpoint;
         }
 
         if ($this->useSSL) {
-            $domain = "https://{$domain}";
+            $scheme = 'https';
         } else {
-            $domain = "http://{$domain}";
+            $scheme = 'http';
         }
 
-        return rtrim($domain, '/') . '/';
+        return rtrim($scheme . '://' . $domain, '/') . '/';
     }
 
     /**
      * Check the endpoint to see if SSL can be used.
+     *
+     * @return $this
      */
-    protected function checkEndpoint()
+    protected function checkEndpoint(): OssAdapter
     {
-        if (0 === strpos($this->endpoint, 'http://')) {
+        if (str_starts_with($this->endpoint, 'http://')) {
             $this->endpoint = substr($this->endpoint, strlen('http://'));
             $this->useSSL = false;
-        } elseif (0 === strpos($this->endpoint, 'https://')) {
+        } elseif (str_starts_with($this->endpoint, 'https://')) {
             $this->endpoint = substr($this->endpoint, strlen('https://'));
             $this->useSSL = true;
         }
+
+        return $this;
     }
 }
