@@ -16,15 +16,15 @@ trait SignatureTrait
     /**
      * gmt.
      *
-     * @param $time
-     *
+     * @param int $time
      * @return string
-     *
-     * @throws \Exception
+     * @throws \DateMalformedStringException
      */
-    public function gmt_iso8601($time)
+    public function gmt_iso8601(int $time): string
     {
         // fix bug https://connect.console.aliyun.com/connect/detail/162632
-        return (new \DateTime(null, new \DateTimeZone('UTC')))->setTimestamp($time)->format('Y-m-d\TH:i:s\Z');
+        return (new \DateTime('', new \DateTimeZone('UTC')))
+            ->setTimestamp($time)
+            ->format('Y-m-d\TH:i:s\Z');
     }
 }
